pipeline/requirements.txt
# Pipeline Backend — Python Dependencies
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1

# Async
asyncio-mqtt==0.16.2
httpx==0.27.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.29

# Task queue
redis==5.0.3
rq==1.16.1

# ML
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
shap==0.45.0
numpy==1.26.4
pandas==2.2.1

# Experiment tracking
mlflow==2.11.0

# Observability
prometheus-client==0.20.0
structlog==24.1.0

# Data validation
rapidfuzz==3.6.2
chardet==5.2.0

# SSE
sse-starlette==2.0.0

# Security
python-multipart==0.0.9

# RAG
chromadb==0.4.24
